Refactored the NumberCard component to use the new animation I created (previous commit)

+++ src/app/modules/gameboard/components/number-card/number-card.component.ts
@@ -1,25 +1,38 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { BoardService } from '../../services/board.service';
-import {fadeInOut} from "../../animations/animations.animation";
-import { BoardMaster } from '../../interfaces/board-master.interface';
+import {cardRemoved} from "../../animations/animations.animation";
 @Component({
   selector: 'number-card',
   templateUrl: './number-card.component.html',
   styleUrls: ['./number-card.component.scss'],
-  animations: fadeInOut
+    host: {
+    '[@cardRemoved]': "displayCard ? 'in' : 'out'",
+    '(@cardRemoved.done)': 'captureDoneEvent($event)',
+  },
+  animations: cardRemoved
 })
+
 export class NumberCardComponent implements OnInit {
   @Input() value!: number;
   @Input() currentTurn?: string;
+  displayCard: boolean  = true
   constructor(private boardService : BoardService) { }
 
   ngOnInit(): void {
   }
 
+  // This function will trigger the animation of removing a card
+  // and then remove it from the service state
   removeCard(): void {
     if(this.currentTurn === "You") {
-      this.boardService.removeCard(this.value);
+      this.displayCard = false;
     }
+  }
+
+  captureDoneEvent(event: any) {
+      if(event.toState === 'out' && this.currentTurn === "You") {
+        this.boardService.removeCard(this.value);
+      }
   }
 
 }
